requirements.txt
Pillow==10.1.0
PySide6==6.6.2
PyMySQL~=1.1.0
tzlocal~=5.2
pytz~=2024.1